requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.1.2
django-searchable-encrypted-fields==0.1.9
gunicorn==20.0.4
psycopg2==2.8.6
pycryptodome==3.9.8
pytz==2020.1
sqlparse==0.4.1
whitenoise==5.2.0
